Update Proxy-skeleton(2025).py
clientSocket.recv(BUFFER_SIZE) is used to read the request data from the client and decode it to get the HTTP request as a string.

Decompose the HTTP request to extract the method, URI, and version, and process the URI to obtain the host name and resource path. Then construct the path to the cache file and try to read the cache file. If the cache hits, the cache content is read and sent directly to the client.

When no cache is found, create a new socket originServerSocket, resolve the host name with socket.gethostByName (), and connect to port 80 of the origin server.

Construct a new HTTP request line based on the client request (e.g. GET/index. HTTP / 1.1 HTML) and request header (including the Host and the Connection: close), after joining together into a complete request packet, through originServerSocket. Sendall () sent out.

The response data from the origin server is received using the loop call recv(), and then the entire response is forwarded to the client using clientSocket.sendall().

After the complete response is obtained, the data is written to the file and saved to the cache directory, so that it can be returned directly from the cache next time.

When the communication is over, close the socket with the origin server, and close the client write channel and the final client socket.